fix(update): update dependencies where possible, use SHA512 signatures
Some packages have been updated to a slightly newer minor version, and
the package-lock.json file has been updated to reflect this. The SHA512
signatures have been added to the package-lock.json file for all packages
that have them available. SHA512 are a more secure hash than SHA1, and
are now the default for npm.