fix(github-workflow): update the publish worklow permissions
The publish workflow needs to be able to publish a Github release and
comment on issues and pull requests. It also needs to be able to use
OIDC tokens for future npm provenance verificiation. This commit
updates the workflow to request the necessary permissions.

Additional changes have been made to the workflow to ensure that the
correct Node.js version is used, that the npm cache is used, and that
the CI environment variable is set. The latter is necessary to ensure
that the `semantic-release` command is run in CI mode.