fix(webpack): update webpack configuration to use terser-webpack-plugin
The terser-webpack-plugin is the successor to the uglifyjs-webpack-plugin
and is the recommended plugin for minifying JavaScript. It is more actively
maintained and has better support for ES6+ syntax.

Update the babel-loader to be more specific about the version of browsers
that we support. This will allow the terser-webpack-plugin to better
optimize the output.

BREAKING CHANGE: The uglifyjs-webpack-plugin is no longer used. The
terser-webpack-plugin is now used instead.